AR JDBC Adapter Fixes Re Migrate/Reset and Reconnect
Fix several db:[reset,create,migrate] issues that were related to poor handling
of date, time, datetime, timestamp.

Fix dropped connection issues, namely the core Rails JDBC adapter has a defect
related to reconnect in the presence of dropped connections; by design it was
supposed to attempt a reconnect exactly once for each .execute request, but as
it would turn out the logic is broken. So we monkey patch the execute logic here
so that the reconnect attempt does occur. Links to related information can be
found here:

- https://github.com/jruby/activerecord-jdbc-adapter/issues/232
- https://github.com/jruby/activerecord-jdbc-adapter/issues/237